Update thousands.go
convert number N to float64

+++ thousands.go
@@ -21,7 +21,7 @@ func Separate(N interface{}, lang ...string) (string, error) {
 
 	}
 
-	n := fmt.Sprintf("%.2f", N)
+	n := fmt.Sprintf("%.2f", float64(N))
 
 	if len(lang) < 1 {
 		lang[0] = "en"
